pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = "我的个人网站"
authors = ["Joker <spaceprivate@163.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = {extras = ["bcrypt"], version = "^5.0.2"}
python-dotenv = "^1.0.1"
django-uuslug = "^2.0.0"
django-compressor = "^4.4"
django-mdeditor = "^0.1.20"
markdown = "^3.6"
markdown-extensions = "^0.0.1"
mysqlclient = "^2.2.4"
django-haystack = "^3.2.1"
jieba = "^0.42.1"
elasticsearch = "^8.13.1"
django-elasticsearch-dsl = "^8.0"
whoosh = "^2.7.4"
user-agents = "^2.2.0"
django-ipware = "^7.0.1"
requests = "^2.32.2"
pygments = "^2.18.0"
markdown-katex = "^202406.1035"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
